refactor interpreter: models in model dir
