+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "osmia"
-version = "1.10.2"
+version = "1.10.3"
 edition = "2021"
 authors = ["Jkutkut"]
