+++ src/model/interpreter/exit_status.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, PartialEq)]
+pub enum ExitStatus {
+	Okay,
+	Break,
+	Continue,
+	False
+}
