+++ src/model/interpreter/mod.rs
@@ -1,0 +1,7 @@
+mod exit_status;
+mod interpreter_value;
+
+pub use exit_status::ExitStatus;
+pub use interpreter_value::InterpreterValue;
+
+pub type InterpreterResult = Result<(ExitStatus, InterpreterValue), String>;
